Rename package to hask3 and drop support for Python 2.
