+++ .gitlab-ci.yml
@@ -42,15 +42,6 @@ check signature of tag:
     - tags
 
 
-run all tests in Python 2.7:
-  stage: baseline test
-  script: tox -e2.7
-  tags:
-    - python
-  except:
-    - tags
-
-
 run all tests in Python 3.6:
   stage: test
   script: tox -e3.6
