+++ hask3/Data/Eq.py
