+++ hask3/Data/Foldable.py
