+++ hask3/Data/Functor.py
