+++ hask3/lang/lazylist.py
