+++ setup.py
@@ -6,15 +6,16 @@ setup(
     version='0.1.1',
     description=des,
     long_description=open('README.rst').read(),
-    author='Bill Murphy',
-    author_email='billpmurphy92@gmail.com',
-    url='https://github.com/billpmurphy/hask',
-    packages=['hask', 'hask.lang', 'hask.Python', 'hask.Data',
-              'hask.Control'],
+    author='Merchise Autrement',
+    author_email='med@merchise.org',
+    url='https://gitlab.merchise.org/merchise/hask',
+    packages=['hask3', 'hask3.lang', 'hask3.Python', 'hask3.Data',
+              'hask3.Control'],
     package_data={'': ['LICENSE', 'README.md']},
     include_package_data=True,
+    python_requires='>=3.6',
     install_requires=[
-        'xoutil>=1.9.4,<2',
+        'xoutil>=1.9.4',
     ],
     license=open('LICENSE').read(),
     zip_safe=False,
