+++ tests/test_maybe.py
@@ -4,13 +4,13 @@ from __future__ import (division as _py3_division,
 
 import unittest
 
-from hask import has_instance
-from hask import Maybe, Just, Nothing
-from hask import Read, Show, Eq, Functor, Applicative
-from hask import Monad, Num, Foldable, Traversable
-from hask import Typeclass, L
+from hask3 import has_instance
+from hask3 import Maybe, Just, Nothing
+from hask3 import Read, Show, Eq, Functor, Applicative
+from hask3 import Monad, Num, Foldable, Traversable
+from hask3 import Typeclass, L
 
-from hask import H, sig, __, t
+from hask3 import H, sig, __, t
 
 
 te = TypeError
@@ -33,7 +33,7 @@ class TestMaybe(unittest.TestCase):
         self.assertFalse(has_instance(Maybe, Traversable))
 
     def test_show(self):
-        from hask.Prelude import show
+        from hask3.Prelude import show
         self.assertEqual("Just(3)", str(Just(3)))
         self.assertEqual("Just(3)", show(Just(3)))
         self.assertEqual("Just('a')", str(Just("a")))
@@ -133,7 +133,7 @@ class TestMaybe(unittest.TestCase):
             Just(1) <= Just(Just(1))
 
     def test_functor(self):
-        from hask.Prelude import id, fmap
+        from hask3.Prelude import id, fmap
         plus1 = (lambda x: x + 1) ** (H/ int >> int)
         toStr = str ** (H/ int >> str)
 
@@ -163,7 +163,7 @@ class TestMaybe(unittest.TestCase):
         def safediv(x, y):
             return Just(x // y) if y != 0 else Nothing
 
-        from hask.Prelude import flip
+        from hask3.Prelude import flip
         s = flip(safediv)
         self.assertEqual(Just(3), Just(9) >> s(3))
         self.assertEqual(Just(1), Just(9) >> s(3) >> s(3))
@@ -180,17 +180,17 @@ class TestMaybe(unittest.TestCase):
         self.assertEqual(Nothing, (Nothing >> s(3)) >> s(3))
         self.assertEqual(Nothing, Nothing >> s_composed)
 
-        from hask.Control.Monad import join, liftM
+        from hask3.Control.Monad import join, liftM
         self.assertEqual(join(Just(Just(1))), Just(1))
         self.assertEqual(join(Just(Nothing)), Nothing)
         self.assertEqual(liftM(__+1, Just(1)), Just(2))
         self.assertEqual(liftM(__+1, Nothing), Nothing)
 
     def test_functions(self):
-        from hask.Data.Maybe import maybe, isJust, isNothing, fromJust
-        from hask.Data.Maybe import fromMaybe
-        from hask.Data.Maybe import listToMaybe, maybeToList, catMaybes
-        from hask.Data.Maybe import mapMaybe
+        from hask3.Data.Maybe import maybe, isJust, isNothing, fromJust
+        from hask3.Data.Maybe import fromMaybe
+        from hask3.Data.Maybe import listToMaybe, maybeToList, catMaybes
+        from hask3.Data.Maybe import mapMaybe
 
         self.assertTrue(isJust(Just(1)))
         self.assertTrue(isJust(Just(Nothing)))
@@ -216,7 +216,7 @@ class TestMaybe(unittest.TestCase):
         self.assertEqual(L[1, 2], catMaybes(L[Just(1), Just(2)]))
         self.assertEqual(L[1, 2], catMaybes(L[Just(1), Nothing, Just(2)]))
 
-        from hask.Prelude import const
+        from hask3.Prelude import const
         self.assertEqual(L[[]], mapMaybe(const(Nothing), L[1, 2]))
         self.assertEqual(L[1, 2], mapMaybe(Just, L[1, 2]))
         self.assertEqual(L[[]], mapMaybe(Just, L[[]]))
