+++ tox.ini
@@ -1,9 +1,8 @@
 [tox]
-envlist = 2.7,3.6
+envlist = 3.6
 
 [testenv]
 basepython =
-     2.7: python2.7
 	 3.6: python3.6
 	 system: python
